refactor: split action steps into separate functions
